chore: address comments

+++ contracts/src/periphery/cert/interfaces/IEigenDACertVerifier.sol
@@ -5,5 +5,5 @@ import {IEigenDACertVerifierBase} from "src/periphery/cert/interfaces/IEigenDACe
 
 interface IEigenDACertVerifier is IEigenDACertVerifierBase {
     /// @notice Returns the certificate version. Used off-chain to identify how to encode a certificate for this CertVerifier.
-    function certVersion() external view returns (uint8);
+    function certVersion() external view returns (uint64);
 }
